requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
marshmallow==3.20.1
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pydantic==2.5.2
email-validator==2.1.0
cryptography==41.0.8
Werkzeug==3.0.1